Add supprot for methods containing underscore
It's hard to tell from reading RFC 2068 if underscores are allowed or
not in the method name. The official HTTP/1.1 methods does not contain
any underscores, but RTC 2326 (the RTSP protocol) which references the
definitions of the HTTP/1.1 RFC does (i.e. GET_PARAMETER and
SET_PARAMETER).

If would make sense ot use this module to parse RTSP headers as well, so
hence support for methods containing underscores is now added.